Code pour faire la figure 5 avec Alias Fractions et Alias Integers (dans preprocess.c)

+++ c/alias_rust.h
@@ -6,11 +6,13 @@
 
 typedef struct {
     VectorInt aliases;
-    double *prob;
+    //double *prob;
+    unsigned int *prob;
     VectorInt small;
     VectorInt large;
     unsigned int n;
-    double weight_sum;
+    //double weight_sum;
+    unsigned int weight_sum;
 } alias_rust_s;
 
 typedef enum {
@@ -23,14 +25,16 @@ typedef enum {
 
 WeightedError weighted_alias_new(
     alias_rust_s *out,
-    const double *weights,
+    //const double *weights,
+    const unsigned int *weights,
     uint32_t n
 );
 
 unsigned int weighted_alias_sample(
     alias_rust_s *dist,
     unsigned int (*rng_index)(unsigned int),
-    double (*rng_weight)(double max)
+    //double (*rng_weight)(double max)
+    unsigned int (*rng_weight)(unsigned int max)
 );
 
 void weighted_alias_free(alias_rust_s *dist);
